feat(action): inital commit to run taskcat
Create a quick initial commit to run taskcat from a Docker container
using GitHub Actions. This commit's goal is to simply see the taskcat
prompt.

Testing is currently being done on the development machine using act
(nektos/act). Eventually act should be included as a pre-commit hook
and/or a step in a CI/CD pipeline, to ensure that the action is
properly tested.

While the AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY are required as
inputs by the action, they are not passed to the Docker container. As
such, taskcat cannot actually deploy CloudFormation templates at this
time.

Associate issue: #1