feat(action): remove action.yaml AWS cred inputs
GitHub Action inputs are prefixed with "INPUT_". For example, when
passing the AWS_ACCESS_KEY_ID, it would appear to the container as
INPUT_AWS_ACCESS_KEY_ID, which taskcat does not recognize.

By removing the inputs from the action.yaml file, the environment
variables are passed as-is, resolving the issue.

Associated issue: #1