test: add test for missing AWS credentials
Add a Java unit test to ensure that the GitHub Action fails when
taskcat throws a "NoCredentialsException", in the event where the
AWS_ACCESS_KEY_ID and AWS_SECRET_ACCESS_KEY environment variables are
not provided. This first test acts as an introduction for unit tests in
this project.

Python was initially used to develop these tests, but Java was
ultimately chosen due to better support for tools like testcontainers
and MockServer, which will be introduced in future commits to mock AWS
calls.