docs: create project README.md
Create the project's README.md file, detailing how to use the
action-taskcat GitHub Action.

While the post-execution script is still being implemented, the
"Managing credentials" section was added, detailing the risk of
accidentally exposing the AWS account ID in the taskcat_outputs logs
and, if that risk is acceptable, information on how to use the
actions/upload-artifact to publish the taskcat_outputs directory as a
build artifact.

Associated issue: ShahradR/action-taskcat#16