test: add test for taskcat success scenario
Add a test to verify the success scenario, where taskcat is provided
with valid AWS credentials and deploys a CloudFormation template.

Because the mock layer has not yet been implemented, we are actually
hitting AWS to run these tests. The AWS credentials are provided to the
test via Maven Surefire parameters passed through the `mvn` command from
the workflow. The credentials themselves are stored as repository
secrets.

We implement the success scenario ahead of the changes to the Docker
container to ensure that the action works both before and after we use
the taskcat/taskcat container instead of the local Dockerfile.

Associated issue: ShahradR/action-taskcat#20