test: add test to print taskcat help
Add a test that prints taskcat's help output by calling taskcat with no
commands. This test is designed to act as a starting point for the
implementation of a generic `commands` input to the GitHub Action,
allowing users the flexibility of running arbitrary tasckat commands,
rather than the hard-coded `taskcat test run`.

Associated issue: ShahradR/action-taskcat#20