fix(action): use entrypoint file to call taskcat
After trying to reduce dependencies and eliminate the need to create a
new container to run taskcat in GitHub workflows, it seems that it might
be easier to simply control the execution from within a script,
bypassing the restrictions set by the GitHub Action's metadata syntax.

This commit re-introduces the entrypoint.sh script and the Dockerfile.
Building on the official taskcat/taskcat container, the script allows
users to pass arbitrary commands to taskcat from GitHub workflows, as
they would when invoking the taskcat from the command line.

Associated issues: ShahradR/action-taskcat#20, ShahradR/action-taskcat#26