test: test action by running in GitHub workflow
Due to validation and implementation differences between act and the
GitHub hosted runners, we cannot solely rely on the JUnit tests to ensure
the correct functionality of the action.

To help catch errors, a job was added to this repository's GitHub
workflow, which runs the example in this project's README file against
the sample CloudFormation template used by the JUnit tests.

While we can only rely on the action's return code to determine if the
execution was successful, it can act as a simple tool to detect errors
that act might have missed.

Associated issue: ShahradR/action-taskcat#20, ShahradR/action-taskcat#26