ci: add CodeQL GitHub workflow
Add support for CodeQL analysis using a workflow to run code scanning on
this repository every Tuesday at 9:00 AM, and help identify security
vulnerabilities and errors in the project's code.

Note that this feature is still in beta. The workflow used in this
project is auto-generated from the "Set up code scanning" wizard on the
project's "Security" tab, with the optional comments deleted. See
https://git.io/JJaKt for more details.

Associated issue: ShahradR/action-taskcat#37