chore(ts): add the ESLint pre-commit hook
Add a pre-commit hook to run ESLint checks when commiting code against
the project repository. The pre-commit hook is a mirror of the ESLint
package, provided by the pre-commit team. It is available on GitHub at
https://git.io/JU3Ev

The additional dependencies specified in the .pre-commit-config.yaml
file match the packages and versions listed in package.json.

Associated issue: ShahradR/git-template#42