ci: add SonarCloud Scan CI workflow job
Add a job in the main CI workflow to scan the repository's code using
SonarCloud, and "detects bugs, vulnerabilities and code smells" that may
have been introduced in commits.

The workflow uses the SonarCloud Action (https://git.io/JUnmZ) to run
the checks. The SONAR_TOKEN secret was also added ahead of pushing this
commit, allowing the GitHub Action to authenticate against the
SonarCloud.