chore: add glob Node module as a dev dependency
Add the glob Node module (https://git.io/v5fMF) as a dev dependency to
this project. This module is used by replace-in-file to match patterns
when searching for files and folders.

While the module is not directly used in the project, we reference glob
in unit tests to mock different paths without hitting the filesystem,
ensuring that the configuration passed to replace-in-file matches log
files in the taskcat_outputs/ directory.

Associated issue: ShahradR/action-taskcat#2