test: generate Istanbul coverage reports
Generate coverage reports using Jest's built-in support for Istanbul
(https://bit.ly/3j3dto2).

This commit also updates the SonarCloud configuration to point to the
test and coverage files, and updates the CI workflow to add the Jest job
as a prerequisite to the SonarCloud analysis, providing the generated
coverage information as an artifact.