chore: set Node package to private
Set the Node package to private, preventing accidental publication of
the code to a public registry. GitHub Actions are ran directly from the
Git repository, and do not need to be published.

This configuration mimicks the actions/typescript-action example
repository (see https://git.io/JUQla).

See https://bit.ly/3iyTkFy for details on the private parameter in the
package.json file.