chore: remove dist/ directory from .gitignore
Remove the dist/ directory from the .gitignore file. The directory is
used by the ncc compiler as an output for the compiled TypeScript
code.

Although the directory will contain machine-generated code, GitHub
Actions expects the compiled code with all its dependencies to be
committed to the repository. See https://git.io/JTml7 for more details.