fix: copy post-entrypoint script to Dockerfile
Copy the dist/index.js post-entrypoint script to the Docker container.
When this action is invoked from other repositories, the runner cannot
find the script and fails with an error.

Integration tests pass because the file is included in this project, and
is automatically copied to the container.

Associated issue: ShahradR/action-taskcat#47