feat: call maskAccountId, publishTaskcatOutputs
Call the maskAccountId and publishTaskcatOutputs functions from the
maskAndPublishTaskcatArtifacts, wiring the TaskcatArtifactManager class
together.

To help with testing, the instantiation and call to the
TaskcatArtifactManager at the bottom of the post-entrypoint.ts script
has been removed. The post-entrypoint will likely need a controller
class that will call the TaskcatArtifactManager, and a dependency
injection solution will need to be used to help with testing. These will
be explored in future commits.