feat: add aws-account-id parameter as action input
Add the aws-account-id parameter as an input to this GitHub Action. The
parameter is set by the aws-actions/configure-aws-credentials action
(see https://git.io/JtZ6H). The AWS account ID can be passed from one
step to the next by following the instructions at https://git.io/JtnN3

BREAKING CHANGE: the new aws-account-id parameter is now required as an
input to this action.