feat: use post-entrypoint.ts as the main script
Because the GitHub Action post-entrypoint creates a new Docker container
on execution, it cannot be used to retrieve the taskcat_outputs log
files generated by the taskcat run (which would have happened in a
separate container instance).

This commit eliminates the entrypoint.sh script, and updates the
post-entrypoint.ts script to call the taskcat process instead.