feat: add action input param to upgrade taskcat
Add the update_taskcat input parameter to the GitHub Action based on
feedback received in ShahradR/action-taskcat#82. When set to true, the
script runs `pip install --upgrade taskcat`, upgrading taskcat to the
latest version before running tests.

The --minimal-output flag was also added to the script to help with
logging issues. See ShahradR/action-taskcat@9111ae4 for more details.

Note that the parameter name uses an underscore instead of a hyphen to
separate "update" and "taskcat". This is because the input is read as an
environment variable, and only ASCII characters + underscores are
allowed.

Associated issue: ShahradR/action-taskcat#82