fix: remove the --minimal-output flag
Based on feedback received in ShahradR/action-taskcat#90, this commit
removes the --minimal-output flag, which is only valid for the
`test run` command.

This will re-introduce reprint output when running tests, which is not
necessarily ordered when printed through GitHub Actions. This will need
to be revisited in future commits.

Associated issue: ShahradR/action-taskcat#90