fix: update the Node.js and npm versions
Update the Dockerfile definition and bump the Node.js and npm versions
from 12.20.1-r0 to 12.21.0-r0, resolving the "unsatisfiable constraints"
error thrown when building the container, as identified by @arniesaha in
ShahradR/action-taskcat#82.

This is a recurrence of the issue identified in January and tracked in
ShahradR/action-taskcat#74.

Associated issue: ShahradR/action-taskcat#74, ShahradR/action-taskcat#82