ci: add workflow to update action major tag
Introduce Actions Tagger to the repository, a "meta-action to keep your
action versions up-to-date" (see https://git.io/JqcC3 for more details).

This workflow is used to update the major version tag and allows
consumers of this GitHub Action to reference a single tag and receive
minor and patch updates.