fix: capture stderr and log it to core.info
Update the stdio logic and add a handler to log the stderr stream from
the taskcat child process separately from stdout.

The logs were previously being piped to process.stdout, which is outside
our control. By handling the output ourselves, we are able to write the
data to the GitHub core library, and eventually do manipulation for
problem matchers.

This commit also adds unit tests to verify the stream behavior, and
configures the @babel/plugin-transform-runtime plugin to enable the use
of Promises, which we use to add a delay between the ChildProcess
instantiation and mock stdout/stderr outputs, and our Jest assertions.
See https://bit.ly/3zfo6fz for more details.