fix: remove aws-account-id from input parameters
Remove the aws-account-id parameter from the mandatory list of input
parameters for this action. This was to be used in a feature to mask the
account ID from taskcat logs before publishing them, to help mask
sensitive data before publishing the log files to GitHub.

That feature is currently unreleased however, and as such, the parameter
should not be required until the implementation is actually available.