fix: set default for update input parameters
The core.getBooleanInput() function reads environment variables to
get values for the different input parameters.

However, in Unix-based operating systems, if an undefined environment
variable is called, an empty string is returned. Because empty strings
aren't allowed when defining a boolean value in the YAML 1.2 spec, the
getBooleanInput() function throws an exception.

This commit verifies if the update_taskcat and update_cfn_lint
parameters have an empty string value: if so, the parameter's value is
set to false, and the exception is suppressed. This gives parity between
the Bash and TypeScript versions of the application, and allows users to
continue using their CI/CD pipelines with no change.

See https://git.io/JKcmg for details on the getBooleanInput() function
implementation.