fix: remove \r, \n characters from EOL
Remove the carriage return (\r) and line feed (\n) characters from
the Python output's end of line.

We leverage the @action/core's info() function to print text to the
console when running this GitHub Action. However, info() automatically
adds \r\n or \n to the end of the string, depending on the OS (see
https://git.io/JKzVA for info()'s implementation). Because the Python
output also contain newline characters, we end up with an extra blank
line every time we print something to the console.

Because the info() function doesn't provide a way to override this
behavior, we work around it by removing \r\n and \n characters from the
end of our line before printing it. It is important to preserve newline
characters in the middle of a string though, since that's sometimes used
by applications, like to render the taskcat ASCII text art, displayed
when invoking the command.