ci: only auto-merge when CI build passes
Update the "Dependabot auto-merge" GitHub workflow to only merge pull
requests if the build passes.

After the check for major/minor version was removed in commit cbf8dfb,
the workflow automatically merges pull requests, regardless of whether
the other workflows have passed or failed.

This commit removes the fetch-metadata step, and adds the
github-action-wait-for-status status condition to the auto-merge step.