fix: update PyYAML, Cython to resolve Docker build
An issue with the PyYAML library causes the container used to run
Taskcat to fail when installing taskcat using pip. This seems to be
caused by a regression introduced in Cython 3.0.0a10.

The issue was identified and resolved by @cormac-yobota and this commit
implements the fix proposed in issue #350.

For more details, see:
- https://bit.ly/3Ayigfi for the issue opened by @cormac-yobota
  describing the problem and the fix in the action-taskcat repository

- https://bit.ly/3CyCCW8 for the issue opened against the PyYAML project,
  describing the regression introduced in Cython 3.0.0a10

Associated issue: #350

Co-authored-by: Cormac Murphy <cormac@yobota.xyz>