ci: split end-to-end tests in a separate workflow
Split the end-to-end tests that run the tagged versions in a separate
GitHub Actions workflow, to ensure that breaking changes don't block
deployments.

The new workflow is also configured to run daily, and will allow us to
catch issues with the action early.

Note that the end-to-end tests that run the untagged versions (i.e.: the
one in the current commit) are still required to create a release, and
will run with every commit.