ci: abandon v1 release support in end-to-end tests
Abandon support for this action's v1 release, considering the action was
broken for an extended period of time, and the fix to resolve the PyYAML
issue highlighted in #350 cannot be easily applied to the older version.

With the abandonned support, the v1 end-to-end tests are removed from
the CI workflow.